backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Anton Ivanov <tombo763a@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "4.0"
djangorestframework = "3.13.1"
drf-yasg = "^1.20.0"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
